Conflicts:
	src/controller/Controller.java
	src/gui/GUI.java
	src/gui/GenomizerView.java
	src/model/Model.java

+++ src/controller/Controller.java
@@ -632,17 +632,24 @@ public class Controller {
             public void mouseClicked(MouseEvent event) {
                 JList list = (JList) event.getSource();
 
+                String specie = "";
+
                 if (list.getModel().getSize() > 0) {
                     int index = list.locationToIndex(event.getPoint());
+                    
                     CheckListItem item = (CheckListItem) list.getModel()
                             .getElementAt(index);
-
-                    item.setSelected(!item.isSelected());
-
-                    GenomeReleaseData[] genome = model
-                            .getSpecieGenomeReleases(item.getSpecie());
-                    view.setGenomeFileList(genome);
-
+                    
+                    if (item.getSpecie().equals(specie) || specie == "") {
+                        
+                        item.setSelected(!item.isSelected());
+                        
+                        GenomeReleaseData[] genome = model
+                                .getSpecieGenomeReleases(item.getSpecie());
+                        view.setGenomeFileList(genome);
+                        
+                    }
+                    
                     list.repaint(list.getCellBounds(index, index));
                 }
             }
@@ -673,7 +680,8 @@ public class Controller {
         public void run() {
             ProcessFeedbackData[] processFeedbackData = model
                     .getProcessFeedback();
-            if (processFeedbackData != null) {
+            if (processFeedbackData != null && processFeedbackData.length > 0) {
+                System.out.println("Feedbackdata received");
                 view.showProcessFeedback(processFeedbackData);
             }
         }
@@ -687,14 +695,36 @@ public class Controller {
 
         @Override
         public void run() {
-            ArrayList<ExperimentData> expData = view
-                    .getSelectedDataInWorkspace();
-            for (ExperimentData data : expData) {
-                for (FileData fileData : data.files) {
-                    model.deleteFileFromExperiment(fileData);
+            int i = 0;
+            if (JOptionPane
+                    .showConfirmDialog(
+                            null,
+                            "Are you sure you want to delete the selected data from the database?",
+                            "Delete from database", JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION) {
+                ArrayList<ExperimentData> expData = view
+                        .getSelectedDataInWorkspace();
+                for (ExperimentData data : expData) {
+                    for (FileData fileData : data.files) {
+                        model.deleteFileFromExperiment(fileData);
+                        i++;
+                    }
+                }
+                expData = view.getSelectedExperimentsInWorkspace();
+                for (ExperimentData data : expData) {
+                    model.deleteExperimentFromDatabase(data);
+                    i++;
                 }
             }
 
+            if (i == 0) {
+                JOptionPane.showMessageDialog(null, "No data was selected",
+                        "Delete error", JOptionPane.ERROR_MESSAGE);
+            } else {
+                JOptionPane.showMessageDialog(null,
+                        "Selected data was removed from database",
+                        "Delete success", JOptionPane.INFORMATION_MESSAGE);
+                view.removeSelectedFromWorkspace();
+            }
         }
     }
